@RestController 注解等同于@Controller @ResponseBody两个注解的合用

+++ boot-jpa/src/main/java/com/llm/Dao/UserRepositoryImpl.java
@@ -58,7 +58,7 @@ public class UserRepositoryImpl implements UserCustomRepository {
         return page;
     }
     private Page<User> readPage(Query dataQuery,Query countQuery,Pageable pageable){
-        dataQuery.setFirstResult(pageable.getOffset());
+        dataQuery.setFirstResult((int)pageable.getOffset());
         dataQuery.setMaxResults(pageable.getPageSize());
         long total=(Long) countQuery.getSingleResult();
         List<User> content=total>(long) pageable.getOffset()?dataQuery.getResultList(): Collections.emptyList();
